Redoing all of the banking code and adding famistudio audio support
