+++ MMC3_128_128.cfg
@@ -21,7 +21,6 @@ MEMORY {
 
 
 #ROM Addresses:
-
     PRG0: start = $8000, size = $2000, file = %O, fill = yes, define = yes;
 	PRG1: start = $8000, size = $2000, file = %O, fill = yes, define = yes;
 	PRG2: start = $8000, size = $2000, file = %O, fill = yes, define = yes;
@@ -34,12 +33,10 @@ MEMORY {
 	PRG9: start = $8000, size = $2000, file = %O, fill = yes, define = yes;
 	PRG10: start = $8000, size = $2000, file = %O, fill = yes, define = yes;
 	PRG11: start = $8000, size = $2000, file = %O, fill = yes, define = yes;
-	PRG12: start = $8000, size = $2000, file = %O, fill = yes, define = yes;
 
     PRG: start = $a000, size = $4000, file = %O ,fill = yes, define = yes;
-	PRG15: start = $e000, size = $2000, file = %O ,fill = yes, define = yes;
 # make sure the reset code is in e000-ffff, see below	
-
+	PRG15: start = $e000, size = $2000, file = %O ,fill = yes, define = yes;
 
 
 # ! 16 Banks of 8K CHR ROM
@@ -62,23 +59,25 @@ SEGMENTS {
     BSS:      load = RAM,            type = bss, define = yes;
     HEAP:     load = RAM,            type = bss, optional = yes;
     ZEROPAGE: load = ZP,             type = zp;
-    ONCE:     load = PRG,            type = ro,  define = yes;
+    ONCE:     load = PRG,            type = ro,  define = yes, optional = yes;
 	
 	XRAM:	  load = WRAM,			 type = bss, define = yes;
 	
-	BANK0:    load = PRG0,           type = ro, define = yes;
-	BANK1:    load = PRG1,           type = ro, define = yes;
-	BANK2:    load = PRG2,           type = ro, define = yes;
-	BANK3:    load = PRG3,           type = ro, define = yes;
-	BANK4:    load = PRG4,           type = ro, define = yes;
-	BANK5:    load = PRG5,           type = ro, define = yes;
-	BANK6:    load = PRG6,           type = ro, define = yes;
-	BANK7:    load = PRG7,           type = ro, define = yes;
-	BANK8:    load = PRG8,           type = ro, define = yes;
-	BANK9:    load = PRG9,           type = ro, define = yes;
-	BANK10:    load = PRG10,           type = ro, define = yes;
-	BANK11:    load = PRG11,           type = ro, define = yes;
-	BANK12:    load = PRG12,           type = ro, define = yes;
+	BANK0:    load = PRG0,           type = ro, define = yes, optional = yes;
+	BANK1:    load = PRG1,           type = ro, define = yes, optional = yes;
+	BANK2:    load = PRG2,           type = ro, define = yes, optional = yes;
+	BANK3:    load = PRG3,           type = ro, define = yes, optional = yes;
+	BANK4:    load = PRG4,           type = ro, define = yes, optional = yes;
+	BANK5:    load = PRG5,           type = ro, define = yes, optional = yes;
+	BANK6:    load = PRG6,           type = ro, define = yes, optional = yes;
+	BANK7:    load = PRG7,           type = ro, define = yes, optional = yes;
+	BANK8:    load = PRG8,           type = ro, define = yes, optional = yes;
+	BANK9:    load = PRG9,           type = ro, define = yes, optional = yes;
+	BANK10:   load = PRG10,           type = ro, define = yes, optional = yes;
+	BANK11:   load = PRG11,           type = ro, define = yes, optional = yes;
+
+	AUDIO:    load = PRG10,           type = ro, define = yes, optional = yes;
+	SAMPLES:  load = PRG15,           type = ro, define = yes, optional = yes;
 
 	STARTUP:  load = PRG15,          type = ro, define = yes;
 	# needs to be in the e000-ffff fixed bank
