+++ MMC3_128_128_DPCM.cfg
@@ -1,0 +1,122 @@
+# NOTE, this assumes you will keep the last bank fixed to c000-ffff, and swap only the first bank 8000-bfff
+
+
+
+MEMORY {
+#RAM Addresses:
+    # Zero page
+    ZP: start = $00, size = $100, type = rw, define = yes;
+	#note, the c compiler + neslib + famitone2 use about 60 zp addresses, I think
+	
+	#note OAM: start = $0200, size = $0100, define = yes;
+	#note, sprites stored here in the RAM
+	
+	RAM: start = $0300, size = $0400, define = yes;
+	#note, I located the c stack at 700-7ff, see below
+	
+	WRAM: start = $6000, size = $2000, define = yes;
+
+#INES Header:
+    HEADER: start = $0, size = $10, file = %O, fill = yes;
+
+
+#ROM Addresses:
+    PRG0:  start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 0;
+	PRG1:  start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 1;
+	PRG2:  start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 2;
+	PRG3:  start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 3;
+	PRG4:  start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 4;
+	PRG5:  start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 5;
+	PRG6:  start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 6;
+	PRG7:  start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 7;
+	PRG8:  start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 8;
+	PRG9:  start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 9;
+	PRG10: start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 10;
+	PRG11: start = $a000, size = $2000, file = %O, fill = yes, define = yes, bank = 11;
+
+# Banks reserved for DPCM by default
+	PRG12: start = $c000, size = $2000, file = %O, fill = yes, define = yes, bank = 12;
+	PRG13: start = $c000, size = $2000, file = %O, fill = yes, define = yes, bank = 13;
+
+    LOW:   start = $8000, size = $2000, file = %O ,fill = yes, define = yes, bank = 14;
+	FIX:   start = $e000, size = $2000, file = %O ,fill = yes, define = yes, bank = 15;
+# make sure the reset code is in e000-ffff, see below	
+
+# ! 16 Banks of 8K CHR ROM
+    CHR:  start = $0000, size = $20000, file = %O, fill = yes;
+}
+
+
+
+
+SEGMENTS {
+    HEADER:   load = HEADER,         type = ro;
+    
+# this needs to be in the fixed bank between e000 and ffff
+    CODE:     load = FIX,            type = ro,  define = yes;
+    INIT:     load = FIX,            type = ro,  define = yes, optional = yes;
+	STARTUP:  load = FIX, 			 type = ro, define = yes;
+	VECTORS:  load = FIX, start = $fffa, type = ro;
+# in this configuration LOWCODE is $8000 (this is a fixed bank but not guaranteed fixed on startup unlike $e000)
+    LOWCODE:  load = LOW,            type = ro;
+
+    RODATA:   load = PRG0,           type = ro,  define = yes;
+    DATA:     load = PRG0, run = RAM,type = rw,  define = yes;
+    CHARS:    load = CHR,            type = rw;
+    BSS:      load = RAM,            type = bss, define = yes;
+    HEAP:     load = RAM,            type = bss, optional = yes;
+    ZEROPAGE: load = ZP,             type = zp;
+    ONCE:     load = PRG0,           type = ro,  define = yes;
+	
+	XRAM:	  load = WRAM,			 type = bss, define = yes;
+	
+	BANK0:    load = PRG0,           type = ro, define = yes;
+	BANK1:    load = PRG1,           type = ro, define = yes;
+	BANK2:    load = PRG2,           type = ro, define = yes;
+	BANK3:    load = PRG3,           type = ro, define = yes;
+	BANK4:    load = PRG4,           type = ro, define = yes;
+	BANK5:    load = PRG5,           type = ro, define = yes;
+	BANK6:    load = PRG6,           type = ro, define = yes;
+	BANK7:    load = PRG7,           type = ro, define = yes;
+	BANK8:    load = PRG8,           type = ro, define = yes;
+	BANK9:    load = PRG9,           type = ro, define = yes;
+	BANK10:   load = PRG10,          type = ro, define = yes;
+
+	AUDIO:    load = PRG11,          type = ro, define = yes;
+	SAMPLES:  load = PRG12, start = $c000, type = ro;
+	SAMPLES1:  load = PRG13, start = $c000, type = ro;
+}
+
+
+
+
+FEATURES {
+    CONDES: segment = INIT,
+        type = constructor,
+        label = __CONSTRUCTOR_TABLE__,
+        count = __CONSTRUCTOR_COUNT__;
+    CONDES: segment = RODATA,
+        type = destructor,
+        label = __DESTRUCTOR_TABLE__,
+        count = __DESTRUCTOR_COUNT__;
+    CONDES: type = interruptor,
+        segment = RODATA,
+        label = __INTERRUPTOR_TABLE__,
+        count = __INTERRUPTOR_COUNT__;
+}
+
+
+
+
+SYMBOLS {
+
+    __STACKSIZE__: type = weak, value = $0100;     # 1 page stack
+	__STACK_START__: type = weak, value = $0700;
+	
+	NES_MAPPER:	  type = weak, value = 4; 			# mapper number, 4 = mmc3, txrom
+	NES_PRG_BANKS:	 type = weak,  value = 8; 			# number of 16K PRG banks, = 128k
+	NES_CHR_BANKS:	  type = weak, value = 16; 			# number of 8K CHR banks, = 128k
+	NES_MIRRORING:	  type = weak, value = 0; 			# 0 horizontal, 1 vertical, 8 four screen, mmc3 can switch mirroring
+	
+}
+
